reuse balanche checking
